move back to long

+++ ModAPI/NexusAPI.cs
@@ -149,7 +149,7 @@ namespace NexusModAPI
         /// <param name="modChannelID"></param>
         /// <param name="targetServer"></param>
         /// <returns></returns>
-        public bool SendModMsgToServer(byte[] data, ushort modChannelID, byte targetServer)
+        public bool SendModMsgToServer(byte[] data, long modChannelID, byte targetServer)
         {
             return Enabled && (bool)sendModMsgToServer(MyTuple.Create(data, modChannelID, targetServer));
         }
@@ -161,7 +161,7 @@ namespace NexusModAPI
         /// <param name="data"></param>
         /// <param name="modChannelID"></param>
         /// <returns></returns>
-        public bool SendModMsgToAllServers(byte[] data, ushort modChannelID)
+        public bool SendModMsgToAllServers(byte[] data, long modChannelID)
         {
             return Enabled && (bool)sendModMsgToAllServers(MyTuple.Create(data, modChannelID));
         }
@@ -386,7 +386,7 @@ namespace NexusModAPI
             public byte toServerID;
 
             [ProtoMember(25)]
-            public ushort targetModMessageID;
+            public long targetModMessageID;
 
             [ProtoMember(30)]
             public byte[] msgData;
